requirements.txt
# Backend dependencies
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
gunicorn==22.0.0
